projects/login_form/server/requirements.txt
Flask 
flask-cors
mysql-connector-python
scikit-learn
numpy
pandas
